# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Beach Volleyball Empire API
  version: 0.1.0
  description: Beach Volleyball Empire API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /facilities:
    get:
      operationId: getFacilities
      tags: [facilities]
      summary: Get all facilities for the team (auto-created at level 1 if missing)
      responses:
        "200":
          description: List of facilities
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Facility"
        "401":
          description: Unauthorized
  /facilities/{type}/upgrade:
    post:
      operationId: upgradeFacility
      tags: [facilities]
      summary: Upgrade a facility to the next level
      parameters:
        - name: type
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Upgraded facility
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Facility"
        "400":
          description: Already at max level or insufficient budget
        "401":
          description: Unauthorized
        "404":
          description: Facility not found
...
  /club-rating:
    get:
      operationId: getClubRating
      tags: [facilities]
      summary: Get the club overall rating with weighted breakdown
      responses:
        "200":
          description: Club rating breakdown
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClubRating"
        "401":
          description: Unauthorized
        "404":
          description: Team not found
...
components:
  schemas:
...
    Facility:
      type: object
      required:
        - id
        - teamId
        - type
        - level
      properties:
        id:
          type: integer
        teamId:
          type: integer
        type:
          type: string
        level:
          type: integer
        upgradingToLevel:
          type: integer
          nullable: true
        upgradeCompletesAtRound:
          type: integer
          nullable: true
        upgradeRoundsRemaining:
          type: integer
          nullable: true
        upgradeBuildLabel:
          type: string
          nullable: true
    ClubRatingComponent:
      type: object
      required: [score, weight, contribution, detail]
      properties:
        score:
          type: integer
        weight:
          type: integer
        contribution:
          type: integer
        detail:
          type: string
    ClubRating:
      type: object
      required: [totalRating, label, breakdown]
      properties:
        totalRating:
          type: integer
        label:
          type: string
        breakdown:
          type: object
          required: [players, staff, medical, facilities, youthAcademy]
          properties:
            players:
              $ref: "#/components/schemas/ClubRatingComponent"
            staff:
              $ref: "#/components/schemas/ClubRatingComponent"
            medical:
              $ref: "#/components/schemas/ClubRatingComponent"
            facilities:
              $ref: "#/components/schemas/ClubRatingComponent"
            youthAcademy:
              $ref: "#/components/schemas/ClubRatingComponent"
...